+++ src/components/Main.tsx
@@ -1,5 +1,5 @@
+import { Box } from '@/components/Box';
 import { Z } from '@/config/style';
-import { Box } from '@components/Box';
 
 interface Props {
   children?: any;
