+++ src/pages/index.tsx
@@ -1,11 +1,11 @@
+import { Banner } from '@/components/Banner';
+import { Main } from '@/components/Main';
+import { Nav } from '@/components/Nav/Nav';
 import { About } from '@/sections/index/About';
 import { Hero } from '@/sections/index/Hero';
 import { Joke } from '@/sections/index/Joke';
 import { Quotes } from '@/sections/index/Quotes';
 import { Technologies } from '@/sections/index/Technologies';
-import { Banner } from '@components/Banner';
-import { Main } from '@components/Main';
-import { Nav } from '@components/Nav/Nav';
 
 export default function Index() {
   return (
