+++ src/sections/index/Joke.tsx
@@ -1,10 +1,10 @@
+import { Section } from '@/components/Section';
+import { BigText } from '@/components/Typography/BigText';
 import { style } from '@/config/style';
 import { rem } from '@/css';
 import { useMediaQuery } from '@/hooks/useMediaQuery';
 import { getJoke } from '@/pages/api/joke';
 import { useEffect, useState } from 'react';
-import { Section } from '@components/Section';
-import { BigText } from '@components/Typography/BigText';
 
 export const Joke: React.FC = () => {
   const [joke, setJoke] = useState<string>(
